+++ geotrek/infrastructure/models.py
@@ -55,6 +55,36 @@ class InfrastructureCondition(StructureOrNoneRelated):
         return self.label
 
 
+class InfrastructureMaintenanceDifficultyLevel(StructureOrNoneRelated):
+    label = models.CharField(verbose_name=_("Label"), max_length=250)
+
+    class Meta:
+        verbose_name = _("Infrastructure Maintenance Difficulty Level")
+        verbose_name_plural = _("Infrastructure Maintenance Difficulty Levels")
+        ordering = ('label',)
+        unique_together = ('label', 'structure')
+
+    def __str__(self):
+        if self.structure:
+            return "{} ({})".format(self.label, self.structure.name)
+        return self.label
+
+
+class InfrastructureUsageDifficultyLevel(StructureOrNoneRelated):
+    label = models.CharField(verbose_name=_("Label"), unique=True, max_length=250)
+
+    class Meta:
+        verbose_name = _("Infrastructure Usage Difficulty Levels")
+        verbose_name_plural = _("Infrastructure Usage Difficulty Levels")
+        ordering = ('label',)
+        unique_together = ('label', 'structure')
+
+    def __str__(self):
+        if self.structure:
+            return "{} ({})".format(self.label, self.structure.name)
+        return self.label
+
+
 class BaseInfrastructure(BasePublishableMixin, Topology, StructureRelated):
     """ A generic infrastructure in the park """
     topo_object = models.OneToOneField(Topology, parent_link=True,
@@ -123,6 +153,18 @@ class Infrastructure(MapEntityMixin, BaseInfrastructure):
     """ An infrastructure in the park, which is not of type SIGNAGE """
     type = models.ForeignKey(InfrastructureType, verbose_name=_("Type"), on_delete=models.CASCADE)
     objects = InfrastructureGISManager()
+    maintenance_difficulty = models.ForeignKey(InfrastructureMaintenanceDifficultyLevel,
+                                               verbose_name=_("Maintenance difficulty"),
+                                               help_text=_("Danger level of maintenance agents' interventions on infrastructure"),
+                                               blank=True, null=True,
+                                               on_delete=models.SET_NULL,
+                                               related_name='infrastructures_set')
+    usage_difficulty = models.ForeignKey(InfrastructureUsageDifficultyLevel,
+                                         verbose_name=_("Usage difficulty"),
+                                         help_text=_("Danger level of end users' infrastructure usage"),
+                                         blank=True, null=True,
+                                         on_delete=models.SET_NULL,
+                                         related_name='infrastructures_set')
 
     class Meta:
         verbose_name = _("Infrastructure")